# openapispec.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  description: Service Checker Service
  title: Service Checker Service
  contact:
    name: Douglas Barthram
servers:
  - url: http://unknown/api
paths:
...
  /service/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema:
          type: string
    delete:
      summary: Remove Service
      description: Remove a service to keep the status for
      operationId: unregister
      responses:
        '204':
          description: Successfully processed
        '404':
          description: Bad request
        '500':
          description: Error processing
...